move mocks and setup out of src
