+++ src/backend/sync/branches/branch.test.ts
@@ -4,8 +4,8 @@ import { branchesNoLastCursor, branchesWithLastCursor } from "../../../common/te
 import { mocked } from "ts-jest/utils";
 import { Subscription } from "../../models";
 import { Application } from "probot";
-import { createWebhookApp } from "../../../common/test-utils/probot";
-import createJob from "../../../common/test-utils/setup/create-job";
+import { createWebhookApp } from "../../../../test-utils/probot";
+import createJob from "../../../../test-utils/setup/create-job";
 import { processInstallation } from "../installation";
 import nock from "nock";
