+++ src/backend/sync/pull-requests/pull-requests.test.ts
@@ -2,9 +2,9 @@
 
 import { mocked } from "ts-jest/utils";
 import { Subscription } from "../../models";
-import { createWebhookApp } from "../../../common/test-utils/probot";
+import { createWebhookApp } from "../../../../test-utils/probot";
 import { Application } from "probot";
-import createJob from "../../../common/test-utils/setup/create-job";
+import createJob from "../../../../test-utils/setup/create-job";
 import { processInstallation } from "../installation";
 import nock from "nock";
 import { RepoSyncState } from "../../models/subscription";
